Halton sampling using LowDiscrepensarySample
I have edited the code that calculates the  prime number  and have reduced one for loop by storing the values directly in the while loop. 
prime variable has the stored the value 1 in the index number which are prime else 0. coprimes variable will store the prime numbers needed as per the number of random variables.